requirements.txt
APScheduler==3.1.0
coverage==4.1
dj-database-url==0.4.0
Django==1.9.6
django-filter==0.13.0
django-nose==1.4.3
djangorestframework==3.3.3
gunicorn==19.4.5
Markdown==2.6.6
nose==1.3.7
psycopg2==2.6.1
pytz==2016.4
whitenoise==2.0.6
filelock==2.0.6
